download link operations

+++ src/main/java/com/example/capstonecckma/repositories/DocRepository.java
@@ -3,6 +3,6 @@ package com.example.capstonecckma.repositories;
 import com.example.capstonecckma.model.Doc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocRepository extends JpaRepository<Doc, Long> {
+public interface DocRepository extends JpaRepository<Doc, Integer> {
 
 }
